Upload
Esta classe é utilizada para capturar eventos e ações feitas pelo usuário
    ou sessão no sistema, e salva-la na forma de log em banco de dados.
    Seu funcionamento e semelhante as triggers, esta classe é eficiente para
    monitorar eventos específicos do lado de dentro do sistema.


This category has been translated by Google. For more information about the product, visit the webpage:
Or not session system, and save it in the form of registration in the database.
Its functioning and the triggers, this class is efficient for
Monitor specific events on the inside of the system.